Move App contents to TaskDashboard container
